YData Profiling for the IMF Data
Performed YData Profiling for examining key aspects to ensure data quality. The analysis includes checking for distinct values, missing entries, negative and positive values, and calculating minimum, maximum, and mean values. It also evaluates the presence of zero and infinite values and assesses the memory usage of the dataset. This comprehensive evaluation is aimed at preparing the data for advanced analytical applications.